Fix a bug of SDL::TTF#draw* when empty string is given
These methods raise error when empty string is given
as a rendered string, but these methods should do nothing in that case.
  * http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=583274
  * http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=714501

+++ lib/sdl.rb
@@ -183,6 +183,7 @@ module SDL
         module_eval(<<-EOS)
           def draw#{t}(dst, text, x, y, #{args})
             image = render#{t}(text, #{args})
+            return if image.nil?
             dst.put(image, x, y)
             image.destroy
           end
